fix script names
