--- client/Assets/Scripts/Music/MainThemeManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0deb2529275b947d4a30259aa1bea285
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
