+++ client/Assets/Scripts/Music/MusicManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 34fed24cf9f254264ba1e42201e10016
+guid: 0deb2529275b947d4a30259aa1bea285
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
